Fix a race on Solaris by pausing when trying to clean locks.
bk: 38aa44bba_9_6NFyGTZtYYagtSnVWQ

+++ src/locking.c
@@ -350,20 +350,27 @@ repository_stale(char *path, int discard)
 	char	*s = strrchr(path, '/');
 	char	host[256];
 	char	*thisHost = sccs_gethost();
-	int	flags = 0;
+	int	slp, tries, flags = 0;
 	u32	pid;
 
 	unless (thisHost) return (0);
 	if (s) s++; else s = path;
 	sscanf(s, "%d@%s", &pid, host);
-	if (streq(host, thisHost) &&
-	    (kill((pid_t)pid, 0) != 0) && (errno == ESRCH)) {
-		if (discard) {
-		    	verbose((stderr, "bk: discarding stale lock %s\n",
-			    path));
-			unlink(path);
+	/* short timeouts to get past races */
+	for (slp = 5000, tries = 0; tries < 8; tries++, slp <<= 1) {
+		if (streq(host, thisHost) &&
+		    (kill((pid_t)pid, 0) != 0) && (errno == ESRCH)) {
+			if (discard && exists(path)) {
+		    		verbose((stderr,
+				    "bk: discarding stale lock %s\n",
+				    path));
+				unlink(path);
+			}
+			return (1);
 		}
-		return (1);
+		unless (tries) fprintf(stderr, "\n");	/* for regressions */
+		fprintf(stderr, "Sleeping %dms for %s\n", slp/1000, path);
+		usleep(slp);
 	}
 	return (0);
 }
